+++ WarpDrive/display/sprite.hpp
@@ -38,9 +38,9 @@ public:
 	bool pixCollision(const Sprite& other) const;
 	bool pixCollision(const SDLImage& other) const;
 	Rectf intersection(const Rectf& box) const; //though this should live in Rect.h, Templates make it dangerous to be there
-	void setTransp(int i = 128) {transpThreshold = i;}
-	void setDrawAlpha(int i);
-    const AnimSheet& Sheet() const { return *sheets[animations[current.X()].first.X().X()];}
+    void setTransp(unsigned int i = 128) {transpThreshold = i;}
+    void setDrawAlpha(unsigned int i);
+    const AnimSheet& Sheet() const { return *sheets[static_cast<unsigned int>(animations[static_cast<unsigned int>(current.X())].first.X().X())];}
 	void setColourKey(const SDLRGBColour& tc = SDLRGBColour(0,255,255));
 
 protected:
